Ignore initial server requests if cached and stale
Staleness is determined by seeing if the Last-Modified value is older than the current request by the length of the refresh rate plus a 10 second grace period.

This check will only run for the initial request to /live_cache_check/1/, as it's assumed that the other urls, which are time based, won't need such aggressive cache busting.

If the cache is stale, this will set the timeStamp variable to the current time and immediately rerun check().